+++ src/realm/util/timestamp_logger.cpp
@@ -7,8 +7,9 @@ using namespace realm;
 using util::TimestampStderrLogger;
 
 
-TimestampStderrLogger::TimestampStderrLogger(Config config)
-    : m_formatter{std::move(config)}
+TimestampStderrLogger::TimestampStderrLogger(Config config, Level level)
+    : Logger(level)
+    , m_formatter{std::move(config)}
 {
 }
 
@@ -16,6 +17,5 @@ TimestampStderrLogger::TimestampStderrLogger(Config config)
 void TimestampStderrLogger::do_log(Logger::Level level, const std::string& message)
 {
     auto now = std::chrono::system_clock::now();
-    std::cerr << m_formatter.format(now) << ": " << get_level_prefix(level) << message << '\n'; // Throws
-    std::cerr.flush();                                                                          // Throws
+    std::cerr << m_formatter.format(now) << ": " << get_level_prefix(level) << message << std::endl; // Throws
 }