+++ test/client/auth.hpp
@@ -63,7 +63,7 @@ public:
         /// not require a thread-safe logger, and it guarantees that all logging
         /// happens either on behalf of the constructor or on behalf of the
         /// invocation of run().
-        util::Logger* logger = nullptr;
+        std::unique_ptr<util::Logger> logger = nullptr;
 
         /// Specifies the maximum number of TCP connections the client can have
         /// to the auth server at any point in time.
@@ -81,7 +81,8 @@ public:
         std::string request_base_path = "/api/client/v2.0";
     };
 
-    util::Logger& logger;
+    std::unique_ptr<util::Logger> m_logger_ptr;
+    util::Logger& m_logger;
 
     /// The client runs in its own thread with an event loop. auth_address and
     /// auth_port specifies the address and port of an username/password
@@ -119,6 +120,8 @@ public:
     ///
     /// This functions is thread-safe.
     void refresh(std::string refresh_token, std::function<RefreshHandler>);
+
+    static std::string make_http_host(bool auth_ssl, const std::string& auth_address, const port_type& auth_port);
 
     util::network::Service& get_service();
     const std::string& get_auth_address();