+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### Breaking changes
 * Updated `logger_factory` in SyncClientConfig to return a `shared_ptr` instead of a `unique_ptr`
+* `util::RootLogger` has been replaced with `util::Logger`
 
 ### Compatibility
 * Fileformat: Generates files with format v22. Reads and automatically upgrade from fileformat v5.